==> Step2_traffic_transform.ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
import openpyxl
import pandas as pd
import requests as rq
from typing import List, Dict
import googlemaps
from tqdm import tqdm
import numpy as np
import copy
from model.get_traffic_intersection import Traffic, get_traffic_table

## 路口編號

取得每一個路口的編號

intersection_list_df = pd.read_excel('DATA/台北市車流量_111路口.xlsx', engine='openpyxl')
intersection_list = list(intersection_list_df['站號'])

intersection_list

# 下載每一個路口的檔案

if False:
    for intersection_id in intersection_list:
        file_name =  intersection_id + '.xlsx'
        url = TRAFFIC_DATA_URL + file_name

        res = rq.get(url)

        with open(f'DATA/traffic/{file_name}', 'wb') as f:
            f.write(res.content)

# 整理路口資料

將分時的車流量整理為「上午_機車」、「上午_汽車」、「上午_卡車」、「下午_機車」、「下午_汽車」、「下午_卡車」

程式碼請參考 [model/get_traffic_intersection](./model/get_traffic_intersection.py)

traffics:List[Traffic] = []
error_intersections = []
for intersection in tqdm(intersection_list):
    try:
        tr = get_traffic_table(f'DATA/traffic/{intersection}.xlsx')
        tr.extract_traffic_data()
        traffics.append(tr)
    except:
        error_intersections.append(intersection)

traffic_df = pd.DataFrame([tr.to_dict for tr in traffics])
traffic_df.to_csv('車流量整理.csv', index = False, encoding='utf-8')

# 下一步

接著，我將下載 7-11, Youbike 的資料。請參考
[下載 7-11, Youbike 資料](./Step3_get_7-11_youbike_data.ipynb)
