==> Step3_get_7-11_youbike_data.ipynb <==
import pandas as pd
from time import sleep
import numpy as np
import requests as rq
from model.function_get_711 import get_towns_name, get_stores_for_town

# 取得大台北地區所有 7-11 資料

先取得所有大台北地區的縣市名稱

taipei_city_districts = get_towns_name('01')
new_taipei_city_districts = get_towns_name('03')

districts = {
    '台北市': taipei_city_districts,
    '新北市': new_taipei_city_districts
}

districts

有台北市與新北市各個區域之後，就可以互叫他們的 API，因為需要縣市名稱與鄉鎮市區的名稱，才會回傳完整的店家列表

store_list = []
for city, towns in districts.items():
    for town in towns:
        print('Getting ' + town + '...')
        store_list += get_stores_for_town(city, town)
        sleep(1)

store_df = pd.DataFrame(store_list)
store_df.X = store_df.X.astype(float) * 1e-6
store_df.Y = store_df.Y.astype(float) * 1e-6

store_df.to_csv('7-11-coordinate.csv', index = False)

# 取得大台北地區所有 Youbike 站點資料

COORD_SOURCE_URL = "https://apis.youbike.com.tw/json/station-yb2.json"

youbike_request = rq.get(COORD_SOURCE_URL)
youbike_df = pd.DataFrame(
    youbike_request.json()
)

篩選出大台北地區的 Youbike 站點

youbike_df = youbike_df[(youbike_df['area_code'] == '00') | (youbike_df['area_code'] == '05')]
youbike_df.to_csv('DATA/youbike2-coordinate.csv', index = False)

# 下一步

接著，我將計算路口在一定範圍內的 7-11, Youbike 數量。請參考
[計算一定範圍內 7-11, Youbike 數量](./Step4_caculate_7-11_youbike_number.ipynb)
