==> Step4_caculate_7-11_youbike_number.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
from model.count_points_within_radius import count_points_within_radius

# 計算 250/500 公尺內的 7-11, Youbike  數量

accidence_location = pd.read_excel('./路口特徵整理_200m.xlsx')
seven_eleven_df = pd.read_csv('DATA/7-11-coordinate.csv')
youbike_df = pd.read_csv('DATA/youbike2-coordinate.csv')

# threshold_distance = 500 #公尺
threshold_distance = 250

accidence_location_corr = accidence_location[['X','Y']].to_numpy()
seven_eleven_location = seven_eleven_df[['X','Y']].to_numpy()
youbike_location = youbike_df[['lng','lat']].to_numpy()

count_711 = count_points_within_radius(accidence_location_corr['X'], accidence_location_corr['Y'], 
                                       seven_eleven_location['lng'], seven_eleven_location['lat'], 
                                       threshold_distance)
count_711

counts_youbike = count_points_within_radius(accidence_location_corr['X'], accidence_location_corr['Y'], 
                                            youbike_location['lng'], youbike_location['lat'], 
                                            threshold_distance)
counts_youbike

accidence_location_corr['count_711'] = count_711

accidence_location_corr['count_youbike'] = counts_youbike

accidence_location

merged_dataset = pd.merge(accidence_location, accidence_location_corr, on=['X', 'Y'], how='left')

# merged_dataset.to_excel('路口特徵with_7-11&youbike.xlsx', index = False)
# merged_dataset.to_excel('路口特徵with_7-11&youbike_200m.xlsx', index = False)

# 下一步

接著，我將轉換事故資料並與路口資料合併。請參考
[轉換事故資料並合併路口資料](./Step5_transform_merge_data.ipynb)
